Move main loop out of cpu crate (#35)
this commit fixes/closes #{35}
moved run_until_alarm function from (cpu/src/lib.rs) to (cpu/src/lib.rs)
