+++ cli/src/main.rs
@@ -1,6 +1,7 @@
 use std::ffi::OsString;
 use std::fs::File;
 use std::fs::OpenOptions;
+use std::time::Duration;
 
 use clap::{App, Arg};
 use tracing::{event, Level};
@@ -9,7 +10,8 @@ use tracing_subscriber::prelude::*;
 use base::prelude::*;
 use cpu::io::{Petr, TapeIterator};
 use cpu::{
-    run_until_alarm, BasicClock, Clock, ControlUnit, MemoryConfiguration, MemoryUnit, ResetMode,
+    Alarm, BasicClock, Clock, ControlUnit, MemoryConfiguration, MemoryUnit, MinimalSleeper,
+    ResetMode,
 };
 
 fn run(
@@ -23,6 +25,80 @@ fn run(
         event!(Level::ERROR, "Execution stopped: {}", e);
     }
     1
+}
+
+fn run_until_alarm(
+    control: &mut ControlUnit,
+    mem: &mut MemoryUnit,
+    clk: &mut BasicClock,
+    multiplier: Option<f64>,
+) -> Result<(), Alarm> {
+    let mut elapsed_ns: u64 = 0;
+    let mut sleeper = MinimalSleeper::new(Duration::from_millis(2));
+    let mut next_hw_poll = clk.now();
+
+    loop {
+        let now = clk.now();
+        if now >= next_hw_poll {
+            // check for I/O alarms, flag changes.
+            event!(Level::TRACE, "polling hardware for updates");
+            match control.poll_hardware(&now) {
+                Ok(Some(next)) => {
+                    next_hw_poll = next;
+                }
+                Ok(None) => {
+                    next_hw_poll = now + Duration::from_micros(5);
+                }
+                Err(e) => {
+                    event!(
+                        Level::INFO,
+                        "Alarm raised during hardware polling after {}ns",
+                        elapsed_ns
+                    );
+                    return Err(e);
+                }
+            }
+        }
+        let in_limbo = match control.fetch_instruction(mem) {
+            Err(e) => {
+                event!(
+                    Level::INFO,
+                    "Alarm raised during instruction fetch after {}ns",
+                    elapsed_ns
+                );
+                return Err(e);
+            }
+            Ok(some_sequence_is_runnable) => !some_sequence_is_runnable,
+        };
+        if in_limbo {
+            // No sequence is active, so there is no CPU instruction
+            // to execute.  Therefore we can only leave the limbo
+            // state in response to a hardware event.  We already know
+            // that we need to check for that at `next_hw_poll`.
+            let interval: Duration = next_hw_poll - now;
+            event!(
+                Level::TRACE,
+                "machine is in limbo, waiting {:?} for a flag to be raised",
+                &interval,
+            );
+            cpu::time_passes(clk, &mut sleeper, &interval, multiplier);
+            continue;
+        }
+        elapsed_ns += match control.execute_instruction(&clk.now(), mem) {
+            Err(e) => {
+                event!(
+                    Level::INFO,
+                    "Alarm raised during instruction execution after {}ns",
+                    elapsed_ns
+                );
+                return Err(e);
+            }
+            Ok(ns) => {
+                cpu::time_passes(clk, &mut sleeper, &Duration::from_nanos(ns), multiplier);
+                ns
+            }
+        };
+    }
 }
 
 #[derive(Debug)]
